+++ HISTORY.rst
@@ -84,6 +84,8 @@
 - Update forum link in the about menu.
 - Only show the settings namespace list link where it is
   relevant.
+- Add support for the fillcolor argument to the rotate
+  transformation.
 
 3.0.1 (2018-07-08)
 =================